+++ firmware/font.c
@@ -172,12 +172,12 @@ struct font* font_load_in_memory(struct font* pf)
     if ( pf->bits_size < 0xFFDB )
     {
         /* pad to 16-bit boundary */
-        fileptr = (unsigned char *)(((int)fileptr + 1) & ~1);
+        fileptr = (unsigned char *)(((long)fileptr + 1) & ~1);
     }
     else
     {
         /* pad to 32-bit boundary*/
-        fileptr = (unsigned char *)(((int)fileptr + 3) & ~3);
+        fileptr = (unsigned char *)(((long)fileptr + 3) & ~3);
     }
 
     if (noffset)
@@ -247,13 +247,13 @@ struct font* font_load_cached(struct font* pf)
     {
         long_offset = 0;
         /* pad to 16-bit boundary */
-        fileptr = (unsigned char *)(((int)fileptr + 1) & ~1);
+        fileptr = (unsigned char *)(((long)fileptr + 1) & ~1);
     }
     else
     {
         long_offset = 1;
         /* pad to 32-bit boundary*/
-        fileptr = (unsigned char *)(((int)fileptr + 3) & ~3);
+        fileptr = (unsigned char *)(((long)fileptr + 3) & ~3);
     }
 
     if (noffset)
